+++ src/Database/InterTable/index.ts
@@ -1,0 +1,4 @@
+export * as AccountRelated from './AccountRelated';
+export * as RepositoryRelated from './RepositoryRelated';
+export * as IssueRelated from './IssueRelated';
+export * as GroupRelated from './GroupRelated';
