+++ src/Database/Table/Account.ts
@@ -1,4 +1,4 @@
-import {Account, Profile} from '../../Class';
+import {Account} from '../../Class';
 import pool from '../Pool';
 import {
     executeTransaction,
@@ -95,32 +95,4 @@ export async function count(account: Readonly<Partial<Account>>): Promise<number
         `SELECT COUNT(*) AS "count" FROM accounts WHERE ${parameterizedStatement}`,
         [...values]);
     return Number.parseInt(rows[0]['count']);
-}
-
-/**
- * @description 为注册操作编写的接口，可以在一个事务内完成账号和账号资料的创建
- * */
-export async function create(account: Readonly<Account>, profile: Readonly<Profile>): Promise<void>
-{
-    const client = await pool.connect();
-    try
-    {
-        await executeTransaction(client, async (client) =>
-        {
-            await client.query(
-                    `INSERT INTO accounts("username", "hash")
-                     VALUES
-                         ($1, $2)`,
-                [account.username, account.hash]);
-            await client.query(
-                    `INSERT INTO profiles("username", "nickname", "email", "avatar")
-                     VALUES
-                         ($1, $2, $3, $4)`,
-                [profile.username, profile.nickname, profile.email, profile.avatar]);
-        });
-    }
-    finally
-    {
-        client.release();
-    }
 }
