+++ src/Database/Table/Issue.ts
@@ -1,43 +1,6 @@
-import {Issue, IssueComment} from '../../Class';
+import {Issue} from '../../Class';
 import pool from '../Pool';
-import {
-    executeTransaction,
-    generateColumnNamesAndValuesArrayAndParameterString,
-    generateParameterizedStatementAndValuesArray,
-} from '../Function';
-
-/**
- * @description 一并插入 Issue 及对应 Comment
- * */
-export async function insertAndReturnId(issue: Readonly<Omit<Issue, 'id'>>, issueComment: Readonly<Omit<IssueComment, 'id' | 'username' | 'belongsTo'>>): Promise<number>
-{
-    const client = await pool.connect();
-    try
-    {
-        const {rows} = await executeTransaction(client, async client =>
-        {
-            const {values: issueValues, columnNames: issueColumnNames, parameterString: issueParameterString} = generateColumnNamesAndValuesArrayAndParameterString(issue);
-            const result = await client.query(`INSERT INTO "issues" (${issueColumnNames}) VALUES (${issueParameterString}) RETURNING "id"`, issueValues);
-
-            const issueId = Number.parseInt(result['rows'][0]['id']);
-            const {username} = issue;
-            const {values: commentValues, columnNames: commentColumnNames, parameterString: commentParameterString} =
-                generateColumnNamesAndValuesArrayAndParameterString(<IssueComment>{
-                    ...issueComment,
-                    id: undefined,
-                    username,
-                    belongsTo: issueId,
-                });
-            await client.query(`INSERT INTO "issue-comments" (${commentColumnNames}) VALUES (${commentParameterString})`, commentValues);
-            return result;
-        });
-        return Number.parseInt(rows[0]['id']);
-    }
-    finally
-    {
-        client.release();
-    }
-}
+import {executeTransaction, generateParameterizedStatementAndValuesArray} from '../Function';
 
 export async function update(issue: Readonly<Partial<Issue>>, primaryKey: Readonly<Pick<Issue, 'id'>>): Promise<void>
 {
