+++ src/Database/index.ts
@@ -1,1 +1,2 @@
 export * from './Table';
+export * from './InterTable';
